[WIP][20449] DocDB: Add Ysql support for DB Cloning
Summary:
The diff adds the required utilities to support DB cloning in case of Ysql databases. It mainly adds the ability to invoke ysql_dump and ysqlsh from within the master leader. This is required in order to:
1- Generate the schema of the clone database as of a point in time (using ysql_dump) on the source database.
2- Modify the output of ysql_dump to change the database name from source_db_name to clone_db_name.
3- Execute the output of 2 using ysqlsh to generate the PG schema of the clone database.

The implementation includes:
- Add `YsqlBinaryRunner` utility class which allows to run any binary against a postgres backend.
- Add `YsqlDumpRunner` to run ysql_dump with utility methods to modify the output script by replacing the database name.
- Add `YsqlshRunner` to run ysqlsh.
- Add `ClonePgSchemaObjects` to be used to support ysql clonning.

Test Plan: Tested manually by cloning a ysql database DB1 to DB2 as of a point in time. Tests will be added when prerquisit diffs land.

Reviewers: asrivastava, hsunder, jhe

Subscribers: ybase, bogdan, slingam, tfoucher

Differential Revision: https://phorge.dev.yugabyte.com/D32465